% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{build_dnorm_dat}
\alias{build_dnorm_dat}
\title{Build a pdf curve for a normal dist}
\usage{
build_dnorm_dat(xmin, xmax, len, pop_mean, pop_sd)
}
\description{
Build a pdf curve for a normal dist
}
