% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{build_dt_dat}
\alias{build_dt_dat}
\title{Build data for a pdf curve for a t-distribution}
\usage{
build_dt_dat(df_sample, ncp_sample, xmin, xmax, len)
}
\arguments{
\item{df_sample}{sample degrees of freedom}

\item{ncp_sample}{non-centrality parameter}

\item{xmin}{lower limit of x values to include}

\item{xmax}{upper limit of x values to include}

\item{len}{how many points to generate?}
}
\description{
Build data for a pdf curve for a t-distribution
}
