% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_one_way_model.R
\name{build_one_way_data}
\alias{build_one_way_data}
\title{Build simulated data with desired properties}
\usage{
build_one_way_data(
  n_obs = 20,
  n_training_pts = 20,
  x_name = "Explanatory",
  y_name = "Response",
  b0 = 1,
  b1 = 1,
  x_lim = c(0, 1),
  error_mean = 0.01,
  error_sd = 10.2,
  seed = NA,
  mod_sd = 1
)
}
\arguments{
\item{n_training_pts}{number of points to simulate for the model residuals}

\item{x_name}{name of the predictor variable}

\item{y_name}{name of the response variable}

\item{b0}{desired intercept parameter}

\item{b1}{desired x1 slope parameter for x1}

\item{error_sd}{}

\item{seed}{random seed for training data}
}
\description{
Build simulated data with desired properties
}
