% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_one_way_model.R
\name{build_one_way_model}
\alias{build_one_way_model}
\title{Create a fitted on-way lm object with desired parameters}
\usage{
build_one_way_model(
  n_training_pts = 20,
  x_name = "Explanatory",
  y_name = "Response",
  b0 = 1,
  b1 = 1,
  seed = NA,
  mod_sd = 1
)
}
\arguments{
\item{n_training_pts}{number of points to simulate for the model residuals}

\item{x_name}{name of the predictor variable}

\item{y_name}{name of the response variable}

\item{b0}{desired intercept parameter}

\item{b1}{desired x1 slope parameter for x1}

\item{seed}{random seed for training data}

\item{desired}{standard deviation of the residuals}
}
\description{
Builds a one-way linear model that you can use for predicting new data values with desired properties.
}
