% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{build_qnorm_dat}
\alias{build_qnorm_dat}
\title{Build data for a quantile curve for a normal dist}
\usage{
build_qnorm_dat(xmin, xmax, len = 100, pop_mean = 0, pop_sd = 1)
}
\arguments{
\item{xmin}{minimum x-value}

\item{xmax}{maximum x-value}

\item{len}{how many points to create for the curve}

\item{pop_mean}{mean for the population}

\item{pop_sd}{standard deviation for the population}
}
\description{
Build data for a quantile curve for a normal dist
}
