% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{build_ribbons}
\alias{build_ribbons}
\title{Build geom_ribbon objects for plotting a continuous distribution function}
\usage{
build_ribbons(
  ribbon_dat,
  fill_lower = rgb(0, 0.3, 0.8, 0.25),
  fill_middle = rgb(0, 0, 0, 0),
  fill_upper = rgb(0, 0.3, 0.8, 0.25),
  col_lower = "black",
  col_middle = "black",
  col_upper = "black"
)
}
\arguments{
\item{ribbon_dat}{data from which to beuld the ribbons}

\item{fill_lower}{fill color for the lower tail}

\item{fill_middle}{fill color for the non-tail part of the pdf}

\item{fill_upper}{fill color for the upper tail}

\item{col_lower}{border color for the lower tail}

\item{col_middle}{border color for the non-tail part}

\item{col_upper}{border color for the upper tail}
}
\description{
Build geom_ribbon objects for plotting a continuous distribution function
}
