% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{build_tail_dat}
\alias{build_tail_dat}
\title{Separate the lower and upper tails for a pdf dataset}
\usage{
build_tail_dat(dat, lower_x = -1.96, upper_x = NULL)
}
\arguments{
\item{dat}{data from which to build the tails}

\item{lower_x}{lower x-limit for the tail}

\item{upper_x}{upper x-limit for the tail}
}
\description{
Separate the lower and upper tails for a pdf dataset
}
