% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_one_way_model.R
\name{build_two_way_data}
\alias{build_two_way_data}
\title{Build simulated data with desired properties}
\usage{
build_two_way_data(
  n_obs = 20,
  n_training_pts = 20,
  x1_name = "x1",
  x2_name = "x2",
  y_name = "Response",
  b0 = 1,
  b1 = 1,
  b2 = 1,
  b3 = 0.5,
  mod_sd = 1,
  x1_lim = c(0, 10),
  x2_lim = c(2000, 3000),
  x1_mean = NULL,
  x2_mean = NULL,
  x1_sd = NULL,
  x2_sd = NULL,
  interaction_term = TRUE,
  error_mean = 0.01,
  error_sd = 0.2,
  seed = NA,
  x1_unif = FALSE,
  x2_unif = FALSE
)
}
\arguments{
\item{n_training_pts}{number of points to simulate for the model residuals}

\item{x1_name}{name of the x1 variable}

\item{x2_name}{name of the x2 variable}

\item{y_name}{name of the response variable}

\item{b0}{desired intercept parameter}

\item{b1}{desired x1 slope parameter for x1}

\item{b2}{desired slope parameter for x2}

\item{b3}{desired interaction parameter}

\item{x1_lim}{= c(0, 10)}

\item{x2_lim}{= c(2000, 3000),}

\item{x1_mean}{= NULL}

\item{x2_mean}{= NULL,}

\item{x1_sd}{= NULL}

\item{x2_sd}{= NULL,}

\item{interaction_term}{whether to include an interaciton term in the model}

\item{error_sd}{}

\item{seed}{random seed for training data}
}
\description{
Build simulated data with desired properties
}
