% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_one_way_model.R
\name{build_two_way_model}
\alias{build_two_way_model}
\title{Parameterize a linear function from a point and slope}
\usage{
build_two_way_model(
  n_training_pts = 20,
  x1_name = "x1",
  x2_name = "x2",
  y_name = "Response",
  b0 = 1,
  b1 = 1,
  b2 = 1,
  b3 = 0.5,
  interaction_term = TRUE,
  seed = NA,
  mod_sd = 1
)
}
\arguments{
\item{n_training_pts}{number of points to simulate for the model residuals}

\item{x1_name}{name of the x1 variable}

\item{x2_name}{name of the x2 variable}

\item{y_name}{name of the response variable}

\item{b0}{desired intercept parameter}

\item{b1}{desired x1 slope parameter for x1}

\item{b2}{desired slope parameter for x2}

\item{b3}{desired interaction parameter}

\item{interaction_term}{whether to include an interaciton term in the model}

\item{seed}{random seed for training data}
}
\description{
Parameterize a linear function from a point and slope
}
