% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_sst.R
\name{calc_ms}
\alias{calc_ms}
\title{Calculate the various mean sums of squares}
\usage{
calc_ms(dat, digits = -1)
}
\arguments{
\item{dat}{data for which to calculate ss}

\item{digits}{how many digits to round for the calculation?}
}
\description{
Calculate the various mean sums of squares
}
