% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_confints.R
\name{ggplot_confints}
\alias{ggplot_confints}
\title{Create simulated confidence intervals}
\usage{
ggplot_confints(n_intervals, n_samples, alpha = 0.05, pop_mean = 0, pop_sd = 1)
}
\arguments{
\item{n_intervals}{The number of intervals to generate.}

\item{n_samples}{The sample size for the confidence intervals.}

\item{alpha}{Confidence level}

\item{pop_mean}{Population}

\item{pop_sd}{population standard deviation}
}
\description{
Simulation of Frequentist confidence intervals of the mean of
normally-distributed data.
}
