% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_density_boxplot.R
\name{means_density_boxplot}
\alias{means_density_boxplot}
\title{Draw a figure showing two distributions via density and boxplots.}
\usage{
means_density_boxplot(
  x1,
  x2,
  xlm,
  ylm,
  col_dens = 1,
  col_1 = 2,
  col_2 = 4,
  col_adj = 0.3,
  lwd_mean = 2,
  lwd_dens = 1
)
}
\description{
Inspired for Chris Sutherland's figures for Intro to
Quantitiative Ecology at UMass Amherst
}
