% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_sst.R
\name{mk_ss_data}
\alias{mk_ss_data}
\title{Create simulated confidence intervals}
\usage{
mk_ss_data(
  n_per_group = 30,
  group_means = c(2, 6, 4),
  group_sd = 0.5,
  cols = NULL,
  pt_alpha = 0.6
)
}
\arguments{
\item{n_per_group}{how many observations in each group}

\item{group_means}{means for the cells (groups)}

\item{group_sd}{sd for the groups}

\item{cols}{colors for the groups}

\item{pt_alpha}{how transparent shoudl the points be?}
}
\value{
a data.table with columns for TODO
}
\description{
Add line segment representation of residuals using base graphics
}
