% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{norm_cdf_plot}
\alias{norm_cdf_plot}
\title{Make a nice example plot for a normal cumulative distribution function}
\usage{
norm_cdf_plot(
  x_1,
  pop_mean = 0,
  pop_sd = 1,
  xmin = NULL,
  xmax = NULL,
  len = 1000,
  fill_cdf = rgb(0, 0.3, 0.8, 0.25),
  digits = 2,
  lty_density = 2,
  x_lab = "x",
  y_lab = "f(x)",
  title_fmt = "x: \%1$s\\nquantile: \%2$s"
)
}
\arguments{
\item{x_1}{x-values to plot}

\item{pop_mean}{mean of the normal distribution to plot}

\item{pop_sd}{standard deviation of the normal distribution to plot}

\item{xmin}{minimum x value for the plot}

\item{xmax}{maximum x value for the plot}

\item{len}{number of points to calculate for the curve}

\item{fill_cdf}{Color to fill the density area}

\item{digits}{the number of siginifcant digits to display in the title}

\item{lty_density}{the line type for the density curve}

\item{x_lab}{label for the x-axis}

\item{y_lab}{label for the y-axis}

\item{title_fmt}{The format string, to be passed to sprintf() for the title}
}
\value{
A ggplot object
}
\description{
Make a nice example plot for a normal cumulative distribution function
}
