% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{norm_quantile_plot}
\alias{norm_quantile_plot}
\title{A nice demonstration plot of a normal quantile function}
\usage{
norm_quantile_plot(
  p_1,
  pop_mean = 0,
  pop_sd = 1,
  xmin = 0.001,
  xmax = 0.999,
  len = 1000,
  fill_cdf = rgb(0, 0.3, 0.8, 0.25),
  digits = 2,
  lty_density = 2,
  y_lab = "x",
  x_lab = "quantile",
  title_fmt = "quantile: \%1$s\\nx: \%2$s\\n"
)
}
\arguments{
\item{p_1}{p}

\item{pop_mean}{mean of the normal distribution to plot}

\item{pop_sd}{standard deviation of the normal distribution to plot}

\item{xmin}{minimum x value for the plot}

\item{xmax}{maximum x value for the plot}

\item{len}{number of points to calculate for the curve}

\item{fill_cdf}{Color to fill background}

\item{digits}{the number of siginifcant digits to display in the title}

\item{lty_density}{the line type for the density curve}

\item{y_lab}{label for the y-axis}

\item{x_lab}{label for the x-axis}

\item{title_fmt}{The format string, to be passed to sprintf() for the title}
}
\value{
A ggplot object
}
\description{
A nice demonstration plot of a normal quantile function
}
