% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Builds a one-way linear model that you can use for predicting new data values with desired properties.}
\usage{
normalize(x, min = 0, max = 1)
}
\arguments{
\item{x}{values to normalize}

\item{x_min}{min value of the normalized output}

\item{x_max}{max value of the normalized output}
}
\description{
Builds a one-way linear model that you can use for predicting new data values with desired properties.
}
