% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_sst.R
\name{plot_anova_ssw}
\alias{plot_anova_ssw}
\title{Create a graphical representation of the within-group sum of squares in an ANOVA style analysis.}
\usage{
plot_anova_ssw(
  dat = NULL,
  seg_col = gray(0.75),
  seg_lwd = 1,
  group_mean_lwd = 2,
  grand_mean_col = "black",
  grand_mean_lwd = 2,
  show_label = TRUE,
  label_fmt = "SSW = \%0.1f",
  label_cex = 1.3,
  digits = 1,
  pt_cex = 1.5,
  pt_lwd = 0.5,
  pch_1 = 16,
  pch_2 = 1,
  seed = 12345
)
}
\arguments{
\item{dat}{data for which to draw.  If null, data with 3 groups will be generated}

\item{seg_col}{color for the vertical line segments}

\item{seg_lwd}{= width for vertical line segments}

\item{group_mean_lwd}{line width for the group mean colors}

\item{grand_mean_col}{color for grand mean horizontal line}

\item{grand_mean_lwd}{line width for the grand mean horizontal line}

\item{show_label}{whether or not to print the SS quantity}

\item{label_fmt}{The text format for the sum of squares label (if )}

\item{digits}{how many digits to round the SS to?}

\item{pt_cex}{size of the points}

\item{pch_1}{plotting character for main point}

\item{pch_2}{plotting character for border point}

\item{seed}{random seed to be used if dat is NULL}
}
\description{
Create a graphical representation of the within-group sum of squares in an ANOVA style analysis.
}
