% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{plot_norm_tails}
\alias{plot_norm_tails}
\title{Plot a normal distribution with one or both tails shaded}
\usage{
plot_norm_tails(
  lower_tail = 0.025,
  upper_tail = 0.925,
  pop_mean = 0,
  pop_sd = 1,
  xmin = NULL,
  xmax = NULL,
  len = 1000,
  fill_cdf_upper = rgb(0, 0.3, 0.8, 0.25),
  fill_cdf_lower = rgb(0, 0.3, 0.8, 0.25),
  fill_middle = rgb(0, 0, 0, 0),
  y_lab = "f(x)",
  x_lab = "x"
)
}
\description{
Plot a normal distribution with one or both tails shaded
}
