% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_way_model.R
\name{plot_one_way_model}
\alias{plot_one_way_model}
\title{Plot a single-predictor linear model highlighting relevant components}
\usage{
plot_one_way_model(
  sim_dat,
  pch_pt = 21,
  cex_pt = 3.1,
  pt_col = adjustcolor("steelblue", 0.3),
  pt_border = 1,
  plot_step = TRUE,
  step_col = "red",
  step_lwd = 0.8,
  step_min = 0.3,
  step_max = 0.7,
  plot_h1 = TRUE,
  h1_col = "darkblue",
  h1_lty = 1,
  h1_lwd = 1,
  plot_h0 = TRUE,
  h0_col = "darkblue",
  h0_lwd = 1,
  h0_lty = 2,
  th = theme_minimal()
)
}
\arguments{
\item{pch_pt}{= 21}

\item{cex_pt}{= 3.1}

\item{pt_col}{= adjustcolor("steelblue", 0.3)}

\item{pt_border}{= 1}

\item{plot_step}{whether to plot the step representing the rise/run components of the slope}

\item{step_col}{colour of the slope step}

\item{step_lwd}{width of the step line}

\item{step_min}{relative x-position of the start of the step}

\item{step_max}{relative x-position of the end of the step'}

\item{plot_h1}{Whether to plot the regression line}

\item{h1_col}{color for the regression line}

\item{h1_lty}{line type for regression line}

\item{h1_lwd}{width for the regression line}

\item{plot_h0}{whether to plot the null hypothesis line}

\item{h0_col}{color for the null hypothesis line}

\item{h0_lwd}{line width for the null hypothesis line}

\item{h0_lty}{line type for the null hypothesis line}

\item{th}{overall plot theme}
}
\description{
Plot a single-predictor linear model highlighting relevant components
}
