% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resids.R
\name{plot_resids}
\alias{plot_resids}
\title{Add line segment representation of residuals using base graphics}
\usage{
plot_resids(x, y_observed, y_expected, lwd = 2, lty = 1, col = "red")
}
\description{
Add line segment representation of residuals using base graphics
}
