% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_dist.R
\name{plot_t_dist}
\alias{plot_t_dist}
\title{Plot one or more t-distributions}
\usage{
plot_t_dist(
  df,
  xlim = c(-3, 3),
  n_x = 500,
  col = rep(1, length(df)),
  xlab = "",
  ylab = "Probability Density",
  main_fmt = "Degrees of Freedom = \%d",
  hl_y = 0.05,
  hl_lty = 3,
  hl_col = 2,
  oma = c(0, 0, 0, 0),
  mar = c(3, 4, 2, 0.5),
  ylim = c(0, 0.4),
  norm = TRUE,
  col_norm = 2,
  lty_norm = 2,
  ...
)
}
\arguments{
\item{df}{either a single value or a vector of degrees of freedom to plot}

\item{xlim}{limits of the x-values to plot}

\item{n_x}{how many points to simulate.  Larger number = smoother curve}

\item{col}{colors for the t curves}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{main_fmt}{main title formatter}

\item{hl_y}{height of the horizontal line to draw}

\item{hl_lty}{type of horizontal line}

\item{hl_col}{color for horizontal line}

\item{oma}{outer margin}

\item{mar}{main figure margins}

\item{ylim}{y-limits for plotting}

\item{norm}{should the standard normal be plotted as a reference?}

\item{col_norm}{color for normal line}

\item{lty_norm}{line style for normal}
}
\description{
Shows the shape of different t-distributions with differing degrees of freedom.
Can also include a plot of the normal as a reference.
}
