% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{plot_t_tails}
\alias{plot_t_tails}
\title{Plot a t distribution with one or both tails shaded}
\usage{
plot_t_tails(
  lower_tail = 0.025,
  upper_tail = 0.925,
  df_sample = 30,
  ncp_sample = 0,
  xmin = NULL,
  xmax = NULL,
  len = 1000,
  fill_lower = rgb(0, 0.3, 0.8, 0.25),
  fill_middle = rgb(0, 0, 0, 0),
  fill_upper = rgb(0, 0.3, 0.8, 0.25),
  y_lab = "f(x)",
  x_lab = "x",
  t_crit = 0.05
)
}
\description{
Plot a t distribution with one or both tails shaded
}
