% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricker_fun.R
\name{ricker_fun}
\alias{ricker_fun}
\title{Ricker function as specified in McGarigal's notes}
\usage{
ricker_fun(x, a, b)
}
\arguments{
\item{x}{x-values}

\item{a}{scales the function}

\item{b}{scales the exponent}
}
\description{
Ricker function as specified in McGarigal's notes
}
