% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_standard_error.R
\name{sample_standard_error}
\alias{sample_standard_error}
\title{Sample standard error of the mean}
\usage{
sample_standard_error(x, pop_sd = NULL)
}
\arguments{
\item{x}{The data for which to calculate the SSE}

\item{pop_sd}{population standard deviation}
}
\description{
Sample standard error of the mean
}
\examples{

# This should usually be around 0.1
sample_standard_error(rnorm(100))
}
