% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seed_pilfer_simulator.R
\name{seed_pilfer_simulator}
\alias{seed_pilfer_simulator}
\title{Simulate seed predation experiment from Bolker 2008, Ch 1.}
\usage{
seed_pilfer_simulator(
  n_simulations,
  n_stations,
  pilfer_rates,
  species_names = NULL
)
}
\arguments{
\item{n_simulations}{How many simulations to perform?}

\item{n_stations}{Numeric (integer) vector.  How many observations for each species?}

\item{pilfer_rates}{Numeric vector.  Pilfer rates for each species.
Each rate must be between 1 and 0 inclusive.}

\item{species_names}{Optional character vector of species names.}
}
\value{
A data.frame object with the results of the simlations.
}
\description{
Simulate seed predation experiment from Bolker 2008, Ch 1.
}
