% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_confidence_intervals_of_mean.R
\name{simulated_confidence_intervals_of_mean}
\alias{simulated_confidence_intervals_of_mean}
\title{Create simulated confidence intervals}
\usage{
simulated_confidence_intervals_of_mean(
  n_samples,
  n_intervals = 1,
  alpha = 0.05,
  pop_mean = 0,
  pop_sd = 1
)
}
\arguments{
\item{n_samples}{The sample size for the confidence intervals.}

\item{n_intervals}{The number of intervals to generate.}

\item{alpha}{Confidence level}

\item{pop_mean}{Population}

\item{pop_sd}{population standard deviation}
}
\description{
Simulation of Frequentist confidence intervals of the mean of
normally-distributed data.
}
\examples{

require(ggplot2)
ggplot(
  simulated_confidence_intervals_of_mean(n_samples = 20, n_intervals = 20),
  aes(x = simulation, colour = in_sample_ci)) +
geom_point(aes(y = sample_mean)) +
geom_segment(aes(xend = simulation, y = lower_t, yend = upper_t)) +
  geom_hline(yintercept = 0, lty = 2)
}
