% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_anova_sst.R
\name{ss_pts}
\alias{ss_pts}
\title{Overlay points for the sums of squares on a base R plot}
\usage{
ss_pts(dat, pt_cex = 1.5, lwd = 0.5, pch_1 = 16, pch_2 = 1)
}
\arguments{
\item{dat}{data for which to draw the points}

\item{pt_cex}{size of the points}

\item{lwd}{width of the point border}

\item{pch_1}{plotting character for main point}

\item{pch_2}{plotting character for border point}
}
\description{
Overlay points for the sums of squares on a base R plot
}
