% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_density_plot.R
\name{gg_norm_conf_int}
\alias{gg_norm_conf_int}
\title{Plot a confidence interval on a normal curve}
\usage{
gg_norm_conf_int(
  alpha = 0.05,
  pop_mean = 0,
  pop_sd = 1,
  xmin = NULL,
  xmax = NULL,
  len = 1000,
  fill_upper = rgb(0, 0.3, 0.8, 0.25),
  fill_lower = rgb(0, 0.3, 0.8, 0.25),
  fill_middle = rgb(0, 0, 0, 0),
  y_lab = "f(x)",
  x_lab = "x",
  lty_v = 2,
  title_fmt = "Interval contains \%.1f\%s of probabiltiy density.",
  arrow_label_fmt = "$\%1$0.2f \\\\pm \%2$0.2f \\\\times \\\\sigma$",
  arrow_size = 0.3,
  digits_label = 0,
  digits_axis = 3,
  x_auto_breaks = -1:1
)
}
\description{
Plot a confidence interval on a normal curve
}
