% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_point_slope.R
\name{line_point_slope}
\alias{line_point_slope}
\title{Calibrate a linear function from a point and slope}
\usage{
line_point_slope(x, x1, y1, slope)
}
\description{
Calibrate a linear function from a point and slope
}
